Handling 404 errors on deleting category

+++ src/routes/deleteCategory.js
@@ -4,6 +4,12 @@ const deleteCategory = (app) => {
   app.delete('/api/category/:id', (req, res) => {
     Category.findByPk(req.params.id).then((category) => {
       const deletedCategory = category;
+      if (category === null) {
+        const message =
+          "La catégorie demandée n'existe pas. Réessayez avec un autre identifiant";
+
+        return res.status(404).json({ message });
+      }
       Category.destroy({
         where: { id: category.id },
       }).then(() => {
